[nits] Refactor example stories
