+++ example/stories/issues/index.js
@@ -1,1 +1,3 @@
+import './53'
+import './59'
 import './63'
